Basic layout for main screen: divided screen into section for each of the four pictures, then a space for guessing
