+++ pubspec.yaml
@@ -20,6 +20,8 @@ dependencies:
   flutter:
     sdk: flutter
   uuid: ^2.0.2
+  dynamic_theme: ^1.0.0
+  flutter_launcher_icons:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
